Add files via upload
Thanks to the help of a a level 2 senior named Gabriel, we successfully made this amazing registration form.  it was an experiment i will never forget, though i didn't code it all by myself, i learned a lot  which is of great importance, and am still learning from this code.